Merge the libcatch2-dev package into catch2
Starting with v2.13.4 catch2 offers parts of its framework as a static
library, with the intention of speeding up the test compilation
time. The previous version of the debian package did ship the static
library as a distinct package named libcatch2-dev, unfortunately this
broke the cmake files shipped with catch2 that expect the library to
be available.

This commit reverse this decision and ships the static library with
the single catch2 package. The main consequence is that catch2 is now
Arch:any rather than Arch:all as it used to be.